+++ src/main/java/one/dfy/bily/api/space/model/repository/SpaceRepository.java
@@ -3,5 +3,6 @@ package one.dfy.bily.api.space.model.repository;
 import one.dfy.bily.api.space.model.Space;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpaceRepository extends JpaRepository<Space,Integer> {
+public interface SpaceRepository extends JpaRepository<Space,Long> {
+
 }
